returnRoomAsInquiry
